+++ README.md
@@ -15,7 +15,7 @@ output adapters (see [Model API](#model-api)). The command line interface is imp
 
 ```shell
 conda env create -f environment.yml
-conda activate perceiver-io
+conda activate model-io
 export PYTHONPATH=.
 ```
 
@@ -121,13 +121,13 @@ python scripts/img_clf.py fit \
 
 ## Model API
 
-The [model](perceiver/model.py) API is based on generic encoder and decoder classes (`PerceiverEncoder` and 
-`PerceiverDecoder`) and task-specific input and output [adapters](perceiver/adapter.py). The following snippet 
+The [model](model/modules.py) API is based on generic encoder and decoder classes (`PerceiverEncoder` and 
+`PerceiverDecoder`) and task-specific input and output [adapters](model/adapter.py). The following snippet 
 shows how they can be used to create an MNIST image classifier, for example:
 
 ```python
-from perceiver.adapter import ImageInputAdapter, ClassificationOutputAdapter
-from perceiver.model import PerceiverIO, PerceiverEncoder, PerceiverDecoder
+from model.adapter import ImageInputAdapter, ClassificationOutputAdapter
+from model.modules import PerceiverIO, PerceiverEncoder, PerceiverDecoder
 
 latent_shape = (32, 128)
